ext4: fix undefined bit shift result in ext4_fill_flex_info
Patchwork-id: 54036
O-Subject: [RHEL6.4 PATCH 2/2] ext4: fix undefined bit shift result in ext4_fill_flex_info
Bugzilla: 809691
CVE: CVE-2012-2100
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Petr Matousek <pmatouse@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

BZ 809691
https://bugzilla.redhat.com/show_bug.cgi?id=809691

BREW 4931551
https://brewweb.devel.redhat.com/taskinfo?taskID=4931551

This commit has not been accepted upstream yet, but it has been posted
to the ext4-list and reviewed.

The result of the bit shift expression in
'1 << sbi->s_log_groups_per_flex' can be undefined in the case that
s_log_groups_per_flex is 31 because the result of the shift is bigger
than INT_MAX. In reality this probably should not cause much problems
since we'll end up with INT_MIN which will then be converted into
'unsigned int' type, but nevertheless according to the ISO C99 the
result is actually undefined.

Fix this by changing the left operand to 'unsigned int' type.

Note that the commit d50f2ab6f050311dbf7b8f5501b25f0bf64a439b already
tried to fix the undefined behaviour, but this was missed.

Thanks to Laszlo Ersek for pointing this out and suggesting the fix.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reported-by: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>