x86, pvclock: Add flag to indicate that a vm was stopped by the host
Patchwork-id: 54074
O-Subject: [RHEL6 kernel PATCH v3 1/8] x86: pvclock: Add flag to indicate that a vm was stopped by the host
Bugzilla: 739151
RH-Acked-by: Marcelo Tosatti <mtosatti@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Eric B Munson <emunson@mgebm.net>

This flag will be used to check if the vm was stopped by the host when a soft
lockup was detected.  The host will set the flag when it stops the guest.  On
resume, the guest will check this flag if a soft lockup is detected and skip
issuing the warning.

Signed-off-by: Eric B Munson <emunson@mgebm.net>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>

(cherry picked from commit eae3ee7d8a7c59cf63441dedf28674889f5fc477)

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>