rtnetlink: Compute and store minimum ifinfo dump size
Patchwork-id: 54090
O-Subject: [PATCH RHEL6.4 1/3][v2] rtnetlink: Compute and store minimum ifinfo dump size
Bugzilla: 845081
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Backport of upstream commit:

commit c7ac8679bec9397afe8918f788cbcef88c38da54
Author: Greg Rose <gregory.v.rose@intel.com>
Date:   Fri Jun 10 01:27:09 2011 +0000

    rtnetlink: Compute and store minimum ifinfo dump size

    The message size allocated for rtnl ifinfo dumps was limited to
    a single page.  This is not enough for additional interface info
    available with devices that support SR-IOV and caused a bug in
    which VF info would not be displayed if more than approximately
    40 VFs were created per interface.

    Implement a new function pointer for the rtnl_register service that will
    calculate the amount of data required for the ifinfo dump and allocate
    enough data to satisfy the request.

v2: The first patch I submitted turned out to be a kABI breaker.  The
movement in struct netlink_callback broke kABI, but it appeared from my
initial testing that putting min_dump_alloc at the end and using the
standard __GENKSYMS__ wrapper was an easy and valid fix for this.
I also had concerns that the addition of IFLA_EXT_MASK in patch 3 would
cause IFLA_MAX to grow and possibly cause problems that are similar to
what we have seen recently with other enum changes.

Because I had concerns about the this fix and the enum change I asked
dzickus for some help.  He was able to look over this patch-set and do
some additional checking to help me verify that this is kABI-safe.
Thanks to dzickus for the help on this.

This will resolve RHBZ 845081.

Signed-off-by: Jarod Wilson <jarod@redhat.com>