ipmi: Add one interface to get more info of low-level IPMI device
Patchwork-id: 54390
O-Subject: [RHEL6.4 PATCH BZ#803653 06/12] IPMI: Add one interface to get more info of low-level IPMI device
Bugzilla: 803653
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Upstream commit: 16f4232ce4d6855361b4eb56262f4a202295c978

The IPMI smi_watcher will be used to catch the IPMI interface as they
come or go.  In order to communicate with the correct IPMI device, it
should be confirmed whether it is what we wanted especially on the
system with multiple IPMI devices. But the new_smi callback function
of smi_watcher provides very limited info(only the interface number
and dev pointer) and there is no detailed info about the low level
interface. For example: which mechansim registers the IPMI
interface(ACPI, PCI, DMI and so on).

This is to add one interface that can get more info of low-level IPMI
device. For example: the ACPI device handle will be returned for the
pnp_acpi IPMI device.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>