acpi_ipmi: Add a default handler for IPMI operation regions
Patchwork-id: 54395
O-Subject: [RHEL6.4 PATCH BZ#803653 10/12] ACPI: Add a default handler for IPMI operation regions
Bugzilla: 803653
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Upstream commit: Submitted, not yet merged

http://lkml.indiana.edu/hypermail/linux/kernel/1209.3/00333.html

The ACPI spec makes it clear that IPMI operation regions should be declared
inside the scope of an IPMI device. Based on the existence of systems for
which this is untrue, it seems likely that alternative implementations exist
that perform some kind of fallback for regions outside the scope of an IPMI
device. Add a callback to the ACPI IPMI driver to glue these operation
regions onto an IPMI device. Behaviour in the case of multiple controllers
may be unpredictable, but there's clearly no way to know the correct answer
in that case.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>