ipmi: Fix KABI issues with acpi_ipmi addition
Patchwork-id: 54398
O-Subject: [RHEL6.4 PATCH BZ#803653 12/12] IPMI: Fix KABI
Bugzilla: 803653
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Half the internal IPMI structures are on the KABI list, so rework chunks
of the previous patch in order to avoid breakage. The only arguably
dangerous part of this is that ipmi_si_get_smi_info() may get called on
an IPMI device that wasn't created by ipmi_si_intf.c, so I've introduced
a token to indicate that the structure was created there.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>