slab: Move kmalloc tracepoint out of inline code
Patchwork-id: 54709
O-Subject: [RHEL6.4 PATCH 1/3] tracing/slab: Move kmalloc tracepoint out of inline code
Bugzilla: 855906
RH-Acked-by: Mikulas Patocka <mpatocka@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

This is a backport of the commit 85beb58.  It is required to avoid
a circular include for Mikulas' patchset introducing a per-cpu semaphore
implementation.  As an added bonus, it reduces the text size of the kernel.
I had to keep the _notrace variants of the calls in the header file, since
they are listed in the whitelist.

This is related to bug 855906.  I've built test kernels on my system,
and verified it does not break the kabi.

commit 85beb5869a4f6abb52a7cf8e01de6fa57e9ee47d
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Wed Nov 24 16:23:34 2010 -0500

    tracing/slab: Move kmalloc tracepoint out of inline code

    The tracepoint for kmalloc is in the slab inlined code which causes
    every instance of kmalloc to have the tracepoint.

    This patch moves the tracepoint out of the inline code to the
    slab C file, which removes a large number of inlined trace
    points.

      objdump -dr vmlinux.slab| grep 'jmpq.*<trace_kmalloc' |wc -l
    213
      objdump -dr vmlinux.slab.patched| grep 'jmpq.*<trace_kmalloc' |wc -l
    1

    This also has a nice impact on size.

       text	   data	    bss	    dec	    hex	filename
    7023060	2121564	2482432	11627056	 b16a30	vmlinux.slab
    6970579	2109772	2482432	11562783	 b06f1f	vmlinux.slab.patched

    Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
    Signed-off-by: Pekka Enberg <penberg@kernel.org>
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>

Signed-off-by: Jarod Wilson <jarod@redhat.com>