e1000e: add device IDs for i218 (Clarkville-LP)
Patchwork-id: 54711
O-Subject: [RHEL6.4 PATCH BZ862764] e1000e: add device IDs for i218 (Clarkville-LP)
Bugzilla: 862764
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=862764

Upstream status:
This is a backport of the following from the linux tree...

commit 16e310ae6ed352c4963b1f2413fcd88fa693eeda
Author: Bruce Allan <bruce.w.allan@intel.com>
Date:   Tue Oct 9 01:11:26 2012 +0000

    e1000e: add device IDs for i218

    i218 is the next-generation LOM that will be available on systems with the
    Lynx Point LP Platform Controller Hub (PCH) chipset from Intel.  This patch
    provides the initial support of those devices.

    Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
    Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>

Brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=4964498

Test status:
The Clarkville-LP NIC is not yet available for testing purposes. So I
simply ran with the patch on a Beaker system that has a Clarkville NIC
(i.e. intel-sharkbay-dh-01.lab.bos), but only as a demonstration that
the new code doesn't break anything.

Signed-off-by: Jarod Wilson <jarod@redhat.com>