iwcm: Reject connect requests if cmid is not in LISTEN state
Patchwork-id: 54927
O-Subject: [Patch RHEL6 031/193] RDMA/iwcm: Reject connect requests if cmid is not in LISTEN state
Bugzilla: 818562 826346 826348 826349
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Steve Wise <swise@opengridcomputing.com>

When destroying a listening cmid, the iwcm first marks the state of
the cmid as DESTROYING, then releases the lock and calls into the
iWARP provider to destroy the endpoint.  Since the cmid is not locked,
its possible for the iWARP provider to pass a connection request event
to the iwcm, which will be silently dropped by the iwcm.  This causes
the iWARP provider to never free up the resources from this connection
because the assumption is the iwcm will accept or reject this connection.

The solution is to reject these connection requests.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>
(cherry picked from commit 3eae7c9f97415ac1f5ab5db4eeb03cde689eb427)

Signed-off-by: Doug Ledford <dledford@redhat.com>