mlx4: Allow dynamic MTU configuration for IB ports
Patchwork-id: 54731
O-Subject: [Patch RHEL6 037/193] mlx4_core: Allow dynamic MTU configuration for IB ports
Bugzilla: 826346 826348 826349
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Or Gerlitz <ogerlitz@mellanox.com>

Set the MTU for IB ports in the driver instead of using the firmware
default of 2KB (the driver defaults to 4KB).  Allow for dynamic mtu
configuration through a new, per-port sysfs entry.

Since there's a dependency between the port MTU and the max number of
HW VLs the port can support, apply a mim/max approach, using a loop
that goes down from the highest possible number of VLs to the lowest,
using the firmware return status to know whether the requested number
of VLs is possible with a given MTU.

For now, as with the dynamic link type change / VPI support, the sysfs
entry to change the mtu is exposed only when NOT running in SR-IOV
mode.  To allow changing the MTU for the master in SR-IOV mode,
primary-function-initiated FLR (Function Level Reset) needs to be
implemented.

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>

(Cherry picked from 096335b)

Signed-off-by: Doug Ledford <dledford@redhat.com>