mlx4: num cores tx rings for every UP
Patchwork-id: 54732
O-Subject: [Patch RHEL6 086/193] net/mlx4_en: num cores tx rings for every UP
Bugzilla: 826346 826348 826349
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Amir Vadai <amirv@mellanox.com>

Change the TX ring scheme such that the number of rings for untagged packets
and for tagged packets (per each of the vlan priorities) is the same, unlike
the current situation where for tagged traffic there's one ring per priority
and for untagged rings as the number of core.

Queue selection is done as follows:

If the mqprio qdisc is operates on the interface, such that the core networking
code invoked the device setup_tc ndo callback, a mapping of skb->priority =>
queue set is forced - for both, tagged and untagged traffic.

Else, the egress map skb->priority =>  User priority is used for tagged traffic, and
all untagged traffic is sent through tx rings of UP 0.

The patch follows the convergence of discussing that issue with John Fastabend
over this thread http://comments.gmane.org/gmane.linux.network/229877

Cc: John Fastabend <john.r.fastabend@intel.com>
Cc: Liran Liss <liranl@mellanox.com>
Signed-off-by: Amir Vadai <amirv@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

(Cherry picked from bc6a474)

Conflicts:
	All files

Amongst other things, we have the problem that the core kernel does not
support the setup_tc netdev op.  Because we have to tiptoe around that
issue we have various conflicts.  In order to allow the driver to have
the ability to run with or without tc queues enabled, I added a module
parameter.  We default to enabled.  Normally this would be called with
all of the queues quiesced and with the rtnl lock held, but in this case
we only enable it during device init which is similarly safe.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>