mlx4: Fix improper use of "port" parameter in mlx4_en_event
Patchwork-id: 54776
O-Subject: [Patch RHEL6 093/193] net/mlx4_en: Fix improper use of "port" parameter in mlx4_en_event
Bugzilla: 826346 826348 826349
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Jack Morgenstein <jackm@dev.mellanox.co.il>

Port is used as an array index before we know if that is proper.

For example, in the catas event case, port is zero; however,
the port index should lie in the range (1..2).

Fix this by using 'port' only in the events where it is of interest.

Test for port out of range in the default (unhandled event) case,
and do not output a message if it is not an ethernet port.

Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Reviewed-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 13bf58b7604d9adfebb8b7c95e6cfb31ec17c699)

Signed-off-by: Doug Ledford <dledford@redhat.com>