cma: Bind to a specific address family
Patchwork-id: 54872
O-Subject: [Patch RHEL6 117/193] RDMA/cma: Bind to a specific address family
Bugzilla: 818562 826346 826348 826349
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Sean Hefty <sean.hefty@intel.com>

The RDMA CM uses a single port space for all associated (tcp, udp,
etc.) port bindings, regardless of the address family that the user
binds to.  The result is that if a user binds to AF_INET, but does not
specify an IP address, the bind will occur for AF_INET6.  This causes
an attempt to bind to the same port using AF_INET6 to fail, and
connection requests to AF_INET6 will match with the AF_INET listener.
Align the behavior with sockets and restrict the bind to AF_INET only.

If a user binds to AF_INET6, we bind the port to AF_INET6 and
AF_INET depending on the value of bindv6only.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>
(cherry picked from commit 5b0ec991c0576c54db75803fbcb0ef5bebfa0828)

Conflicts:
	drivers/infiniband/core/cma.c

We don't have IS_ENABLED, revert to the if defined() form instead.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>