amd_iommu: attach device fails on the last pci device
Patchwork-id: 54963
O-Subject: [RHEL6.4 PATCH] amd_iommu: attach device fails on the last pci device
Bugzilla: 861164
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

amd_iommu_attach_device() checks if device id is within
the limits of amd_iommu_last_bdf and instead checking if
devid > amd_iommu_last_bdf, it checks devid >= amd_iommu_last_bdf.
As a result the last device attach fails because
amd_iommu_attach_device() returns an -EINVAL.
This bug is in linux-2.6.32 and an equivalent fix in linux-2.6.33
and has been carried forward to later kernels and is in the
upstream kernel. This equivalent fix includes restructuring and
consolidating device checks into a routine
check_device(). Instead of back-porting all of that work, spot-fixed the
bug in amd_iommu_attach_device() for linux-2.6.32.

Above is from upstream posting: https://lkml.org/lkml/2012/10/12/273

Verified by partner who is also the one that posted the patch.
Obvious fix that follows the logic elsewhere in that file
with respect to the use of amd_iommu_last_bdf.

Posting for immediate review & inclusion for RHEL6.4.

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=861164

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4970913

Testing:
Partner tested brew built kernel and verified the RHEL kernel
with this backport fixed his configuration, as the patch did for
a stable 2.6.32 kernel.

Notes: the patch will not be applied to upstream since the
       code was refactored to use a common check-device() function
       for all instances, and is correct in upstream.

Please review & ack.

Signed-off-by: Jarod Wilson <jarod@redhat.com>