kabi: Adding symbol cancel_dirty_page
Patchwork-id: 55066
O-Subject: [PATCH RHEL6.4 BZ866427 03/35] kabi: Adding symbol cancel_dirty_page
Bugzilla: 866427
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Bugzilla: 866427
https://bugzilla.redhat.com/show_bug.cgi?id=866427

Description:
============
This function is used to cancel the dirty bit on kernel page. If there
is any error when allocating a page then that page needs to be
destroyed. It could be posible that between creation and destruction of
the page it could be dirtied, so we need to clear dirty state before
destroying

Tested:
=======
brewed kernel properly