audit: do not print error when SELinux disabled
Patchwork-id: 55126
O-Subject: [RHEL6] Audit: do not print error when SELinux disabled
Bugzilla: 785936
RH-Acked-by: Paul Moore <pmoore@redhat.com>

RHBZ: 785936
About to be posted upstream

If the audit system collects a record about one process sending a signal
to another process it includes in that collection the 'secid' or 'an int
used to represet an SELinux label.'  If SELinux is disabled it will
collect a 0.  The problem is that when we attempt to print that record
we ask the LSM to convert the secid back to a string.  Since there is no
LSM it returns EOPNOTSUPP.

Most code in the audit system checks if the secid is 0 and does not
print LSM info in that case.  The signal information code however forgot
that check.  Thus users will see a message in syslog indicating that
converting the sid to string failed.  Add the right check.

Signed-off-by: Eric Paris <eparis@redhat.com>