fix bug that causes no RX queue entries to be created under /sys/devices
Patchwork-id: 55170
O-Subject: [RHEL6.4 PATCH] net: fix bug that causes no RX queue entries to be created under /sys/devices
Bugzilla: 869770
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Bugzilla: 869770
https://bugzilla.redhat.com/show_bug.cgi?id=869770

Description:
The following backport was not complete and it causes a regression that no RX
queue entries are created under /sys/devices.

[net] Allow changing number of RX queues after device allocation (Neil Horman) [813086]
http://post-office.corp.redhat.com/archives/rhkernel-list/2012-September/msg00436.html

The patch introduced the code with "#ifdef CONFIG_RPS". However, CONFIG_RPS
was not defined in the patch although it is defined in the upstream kernel.
Also, some compile errors were found when CONFIG_RPS was on due to the
difference of the netdevice structure between upstream and RHEL kernels.

This patch does the following to fix the bug:
  - Define CONFIG_RPS
  - Fix the compile errors when CONFIG_RPS is on

Upstream:
N/A

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=5021889

Test:
Tested on multiple systems and verified that RX queue entries were created
under /sys/devices.

Signed-off-by: Jarod Wilson <jarod@redhat.com>