fscache: Provide proper invalidation
Patchwork-id: 55224
O-Subject: [RHEL6 PATCH 12/15] FS-Cache: Provide proper invalidation
Bugzilla: 699931
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Provide a proper invalidation method rather than relying on the netfs retiring
the cookie it has and getting a new one.  The problem with this is that isn't
easy for the netfs to make sure that it has completed/cancelled all its
outstanding storage and retrieval operations on the cookie it is retiring.

Instead, have the cache provide an invalidation method that will cancel or wait
for all currently outstanding operations before invalidating the cache, and
will cause new operations to queue up behind that.  Whilst invalidation is in
progress, some requests will be rejected until the cache can stack a barrier on
the operation queue to cause new operations to be deferred behind it.

Signed-off-by: David Howells <dhowells@redhat.com>