huge_memory: create sysfs symlink to redhat_transparent_hugepage
Patchwork-id: 55244
O-Subject: [RHEL6.4 PATCH v2] create a symlink to redhat_transparent_hugepage
Bugzilla: 852802
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Dave Young <dyoung@redhat.com>

BZ#852802
https://bugzilla.redhat.com/show_bug.cgi?id=852802

On upstream, under/sys/kernel/mm/, the sysfs dir entry for transhuge
is transparent_hugepage. But on RHEL6, the name of it is
redhat_transparent_hugepage.For consistency, create a symlink to
redhat_transparent_hugepage, named as transparent_hugepage.

On i686, the symlink may be conditional on CONFIG_TRANSPARENT_HUGEPAGE.
So it only gets put in place when the redhat_transparent_hugepage
directory actually exists.

And the local static variable is meaningless in function hugepage_init,
the static is removed.

Brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=5036670

Test:
Passed on X86_64.

Signed-off-by: Jarod Wilson <jarod@redhat.com>