ptrace-utrace: fix PTRACE_GETEVENTMSG(pid) in sub-namespace
Patchwork-id: 55248
O-Subject: [RHEL6 PATCH BZ 782330] ptrace-utrace: fix PTRACE_GETEVENTMSG(pid) in sub-namespace
Bugzilla: 782330
RH-Acked-by: Dave Anderson <anderson@redhat.com>
RH-Acked-by: Anton Arapov <Anton@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=782330

ptrace_report_clone() uses child->pid, this is obviously wrong unless
the tracer is from the global namespace. Change it to use task_pid_vnr().

This is still not right, we should use the tracer's namespace, not
parent's. But this matches upstream, and at least this works if they
are from the same namespace.

See also
http://lists.fedoraproject.org/pipermail/kernel/2011-August/003340.html
the problem and the fix are the same.

Reported-by: Robin Green <greenrd@greenrd.org>
Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>