dm-raid: Transiently failed devices do not undergo resync
Patchwork-id: 55255
O-Subject: [PATCH] DM RAID: Transiently failed devices do not undergo resync
Bugzilla: 870246
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

DM RAID devices are not being resync'ed when they are reintroduced into
the array after having been failed for some period of time.  This
problem was addressed upstream (albeit to address a different problem at
the time) in the upstream commit, 81f382f9.  This patch was then pulled
into the RHEL6 kernel in commit 645b6180.

Later updates to MD in RHEL6 commit 8e96078f reversed most of the above
patch that iterates over the devices in an array and sets which ones are
faulty - a necessary condition for their later resync'ing.  I am,
therefore, reintroducing (most of) upstream patch 81f382f9 to align the
code once again with the upstream version - fixing the problem.

Bug 870246:
 https://bugzilla.redhat.com/show_bug.cgi?id=870246

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5050433

Testing:
 Without the patch, device failures are not properly recorded.  When a
device is re-introduced, it is not synced.  After the patch, the
reentering devices are sync'ed.

Upstream status:
 Part of upstream commit 81f382f9.  (See above for more detailed
explanation.)

commit 81f382f9e0b25ef56b1c0283c900b86b91a5e4c7
Author: Jonathan Brassow <jbrassow@redhat.com>
Date:   Tue May 22 13:55:30 2012 +1000

    DM RAID: Record and handle missing devices

    Missing dm-raid devices should be recorded in the superblock

    When specifying the devices that compose a DM RAID array, it is possible to denote
    failed or missing devices with '-'s.  When this occurs, we must record this in the
    superblock.  We do this by checking if the array position's data device is missing
    and then forcing MD to record the superblock by setting 'MD_CHANGE_DEVS' in
    'raid_resume'.  If we do not cause the superblock to be rewritten by the resume
    function, it is possible for a stale superblock to be written by an out-going
    in-active table (during 'raid_dtr').

    Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
    Signed-off-by: NeilBrown <neilb@suse.de>

Signed-off-by: Jarod Wilson <jarod@redhat.com>